+++ data-raw/regiones_y_comunas.R
@@ -1,5 +1,5 @@
 ## code to prepare `regiones_y_comunas` dataset goes here
 
-regiones_y_comunas <- read.csv2('data-raw/regiones_y_comunas.csv')
+regiones_y_comunas <- read.csv2('data-raw/regiones_y_comunas.csv', encoding = 'UTF-8')
 
 usethis::use_data(regiones_y_comunas, overwrite = TRUE)