+++ NEWS.md
@@ -10,11 +10,13 @@
 ## Features
 
 * Agregar columna `comuna18` para tener el código comunal luego de la creación de la región del Ñuble en 2018.
+* Nuevas variables en `regiones_y_comunas` para ajustes de no respuesta.
 * Base de datos con `codigos_ensenanza` para tener equivalencias y descripción de la nomenclatura utilizada por MINEDUC en sus bases de datos.
 
 ### Gráfico
 
 * Función `gg_bar_3_niveles_stack`.
+
 
 ## Bug fix